refactor(dashboard): split AssistantBlocks into per-block-type files
AssistantBlocks.tsx had grown to 1,070 lines containing every block
renderer plus the streaming indicator, the activity grouper, and the
top-level dispatcher. Each renderer is now its own file under
chat/blocks/, leaving AssistantBlocks.tsx as a 97-line dispatcher
(BlockSegmentView) + root component.

New files:
- blocks/format-tool-args.ts (shared utility for tool arg previews)
- blocks/StreamingIndicator.tsx (NeuralOrganism + activity bars)
- blocks/ThinkingBlockView.tsx
- blocks/StatusBlockView.tsx
- blocks/LogOutputView.tsx
- blocks/TextBlockView.tsx (markdown + syntax highlighting + packed badges)
- blocks/ToolUseBlockView.tsx (tool result + emit_interaction buttons)
- blocks/AgentBlockView.tsx (skill/subagent + advise/findings/impact views)
- blocks/TodoBlockView.tsx
- blocks/ActivityGroup.tsx (tool clustering + mixed-block grouping)

External API unchanged (AssistantBlocks, BlockSegmentView still exported
from chat/AssistantBlocks). Full dashboard suite (315 tests) passes.
