+++ packages/dashboard/src/client/components/chat/blocks/ThinkingBlockView.tsx
@@ -1,0 +1,17 @@
+import type { ThinkingBlock } from '../../../types/chat';
+
+export function ThinkingBlockView({ block }: { block: ThinkingBlock }) {
+  return (
+    <details className="rounded border border-neutral-border/50 bg-neutral-surface/50 backdrop-blur-sm group">
+      <summary className="cursor-pointer px-3 py-1.5 text-xs font-medium text-neutral-muted select-none flex items-center gap-2">
+        <div className="h-1 w-1 rounded-full bg-secondary-400 group-open:animate-pulse" />
+        Thinking...
+      </summary>
+      <div className="border-t border-neutral-border/50 px-3 py-2">
+        <p className="whitespace-pre-wrap text-xs leading-relaxed text-neutral-muted">
+          {block.text}
+        </p>
+      </div>
+    </details>
+  );
+}
